+++ library/common/Enumerations.php
@@ -10,6 +10,7 @@ abstract class PaymentType
 {
     const Immediate = "Immediate";
     const Reservation = "Reservation";
+    const DelayedCapture = "DelayedCapture";
 }
 
 abstract class FundingSourceType
@@ -17,6 +18,7 @@ abstract class FundingSourceType
     const All = "All";
     const Balance = "Balance";
     const Bankcard = "Bankcard";
+    const BankTransfer = "BankTransfer";
 }
 
 abstract class PaymentStatus
@@ -27,8 +29,12 @@ abstract class PaymentStatus
     const Started = "Started";
     // 21
     const InProgress = "InProgress";
+    // 22
+    const Waiting = "Waiting";
     // 25
     const Reserved = "Reserved";
+    // 26
+    const Authorized = "Authorized";
     // 30
     const Canceled = "Canceled";
     // 40
@@ -91,4 +97,112 @@ abstract class CardType
     const Visa = "Visa";
     const Electron = "Electron";
     const AmericanExpress = "AmericanExpress";
+}
+
+
+// --------------------
+// 3D Secure properties
+// --------------------
+
+abstract class AccountCreationIndicator
+{
+    const NoAccount = "NoAccount";
+    const CreatedDuringThisTransaction = "CreatedDuringThisTransaction";
+    const LessThan30Days = "LessThan30Days";
+    const Between30And60Days = "Between30And60Days";
+    const MoreThan60Days = "MoreThan60Days";
+}
+
+abstract class AccountChangeIndicator
+{
+    const ChangedDuringThisTransaction = "ChangedDuringThisTransaction";
+    const LessThan30Days = "LessThan30Days";
+    const Between30And60Days = "Between30And60Days";
+    const MoreThan60Days = "MoreThan60Days";
+}
+
+abstract class PasswordChangeIndicator
+{
+    const NoChange = "NoChange";
+    const ChangedDuringThisTransaction = "ChangedDuringThisTransaction";
+    const LessThan30Days = "LessThan30Days";
+    const Between30And60Days = "Between30And60Days";
+    const MoreThan60Days = "MoreThan60Days";
+}
+
+abstract class ShippingAddressUsageIndicator
+{
+    const ThisTransaction = "ThisTransaction";
+    const LessThan30Days = "LessThan30Days";
+    const Between30And60Days = "Between30And60Days";
+    const MoreThan60Days = "MoreThan60Days";
+}
+
+abstract class PaymentMethodIndicator
+{
+    const NoAccount = "NoAccount";
+    const ThisTransaction = "ThisTransaction";
+    const LessThan30Days = "LessThan30Days";
+    const Between30And60Days = "Between30And60Days";
+    const MoreThan60Days = "MoreThan60Days";
+}
+
+abstract class SuspiciousActivityIndicator
+{
+    const NoSuspiciousActivityObserved = "NoSuspiciousActivityObserved";
+    const SuspiciousActivityObserved = "SuspiciousActivityObserved";
+}
+
+abstract class DeliveryTimeframeType
+{
+    const ElectronicDelivery = "ElectronicDelivery";
+    const SameDayShipping = "SameDayShipping";
+    const OvernightShipping = "OvernightShipping";
+    const TwoDayOrMoreShipping = "TwoDayOrMoreShipping";
+}
+
+abstract class AvailabilityIndicator
+{
+    const MerchandiseAvailable = "MerchandiseAvailable";
+    const FutureAvailability = "FutureAvailability";
+}
+
+abstract class ReOrderIndicator
+{
+    const FirstTimeOrdered = "FirstTimeOrdered";
+    const ReOrdered = "ReOrdered";
+}
+
+abstract class ShippingAddressIndicator
+{
+    const ShipToCardholdersBillingAddress = "ShipToCardholdersBillingAddress";
+    const ShipToAnotherVerifiedAddress = "ShipToAnotherVerifiedAddress";
+    const ShipToDifferentAddress = "ShipToDifferentAddress";
+    const ShipToStore = "ShipToStore";
+    const DigitalGoods = "DigitalGoods";
+    const TravelAndEventTickets = "TravelAndEventTickets";
+    const Other = "Other";
+}
+
+abstract class PurchaseType
+{
+    const GoodsAndServicePurchase = "GoodsAndServicePurchase";
+    const CheckAcceptance = "CheckAcceptance";
+    const AccountFunding = "AccountFunding";
+    const QuasiCashTransaction = "QuasiCashTransaction";
+    const PrePaidVacationAndLoan = "PrePaidVacationAndLoan";
+}
+
+abstract class RecurrenceType 
+{
+    const MerchantInitiatedPayment = "MerchantInitiatedPayment";
+    const OneClickPayment = "OneClickPayment";
+    const RecurringPayment = "RecurringPayment";
+}
+
+abstract class ChallengePreference 
+{
+    const NoPreference = "NoPreference";
+    const ChallengeRequired = "ChallengeRequired";
+    const NoChallengeNeeded = "NoChallengeNeeded";
 }